+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.0.2-alpha.1",
-  "contentSha256_12": "9e0f1f39b94e",
+  "pkgVersion": "1.0.2-alpha.2",
+  "contentSha256_12": "8c2da84119af",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.0.2-alpha.1";
-export const contentSha256_12 = "9e0f1f39b94e";
+export const pkgVersion = "1.0.2-alpha.2";
+export const contentSha256_12 = "8c2da84119af";
 export const file = "nodalia-cards.bundle.js";